Fixed a bug

+++ SpaceTrader/app/src/main/java/onenine/android/Model/Goods.java
@@ -52,7 +52,7 @@ public enum Goods {
     }
 
     private double calculateVar() {
-        int randVar = new Random().nextInt(this.var);
+        double randVar = new Random().nextInt(this.var);
         return randVar/100;
     }
